--- apis/MyDriving.POIService.v2/Dockerfile
@@ -1,6 +1,0 @@
-#FROM microsoft/azure-functions-runtime:2.0.0-jessie
-
-FROM azurefunctionstest/azure-functions-dotnet2.0:dev-jessie
-
-ENV AzureWebJobsScriptRoot=/home/site/wwwroot
-COPY ./bin/ /home/site/wwwroot
