--- apis/MyDriving.POIService.v2/HealthCheckService.cs
@@ -1,22 +1,0 @@
-
-using System.IO;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Azure.WebJobs.Host;
-using Newtonsoft.Json;
-
-namespace MyDriving.POIService.v2
-{
-    public static class HealthCheckService
-    {
-        [FunctionName("HealthCheck")]
-        public static IActionResult Run([HttpTrigger(AuthorizationLevel.Function, "get", "post", Route = null)]HttpRequest req, TraceWriter log)
-        {
-            log.Info("C# HTTP trigger function processed a request.");
-
-            return (ActionResult)new OkObjectResult("POI Service is healthy.");
-        }
-    }
-}
