--- apis/MyDriving.POIService.v2/MyDriving.POIService.v2.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2036
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyDriving.POIService.v2", "MyDriving.POIService.v2.csproj", "{DFE372D1-13E9-4CBC-8A4B-974455AAF7A6}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MyDriving.ServiceObjects", "..\MyDriving.ServiceObjects\MyDriving.ServiceObjects.csproj", "{8504596A-A3F0-40D2-91D6-B1630D1051E1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DFE372D1-13E9-4CBC-8A4B-974455AAF7A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DFE372D1-13E9-4CBC-8A4B-974455AAF7A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DFE372D1-13E9-4CBC-8A4B-974455AAF7A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DFE372D1-13E9-4CBC-8A4B-974455AAF7A6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8504596A-A3F0-40D2-91D6-B1630D1051E1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8504596A-A3F0-40D2-91D6-B1630D1051E1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8504596A-A3F0-40D2-91D6-B1630D1051E1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8504596A-A3F0-40D2-91D6-B1630D1051E1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E186AD42-743A-492B-A340-ECB61DBE4395}
-	EndGlobalSection
-EndGlobal
