--- apis/MyDriving.POIService.v2/appSettings.json
@@ -1,6 +1,0 @@
-﻿{
-  "ConnectionStrings": {
-    //"myDrivingDB": "Server=tcp:mydrivingdbserver-vpwupcazgfita.database.windows.net,1433;Initial Catalog=mydrivingDB;Persist Security Info=False;User ID=YourUserName;Password=MyComplex-Passw0rd;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;",
-    "myDrivingDB": "Server=tcp:[SQL_SERVER],1433;Initial Catalog=[SQL_DBNAME];Persist Security Info=False;User ID=[SQL_USER];Password=[SQL_PASSWORD];MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
-  }
-}
