--- apis/MyDriving.POIService.v2/helm/templates/deployment.yaml
@@ -1,50 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: "{{ .Values.image.name }}-deploy"
-  labels:
-    deploy: {{ .Values.image.label }}
-spec:
-  replicas: {{ .Values.replicaCount }}
-  selector:
-    matchLabels:
-      app: {{ .Values.image.label }}
-  template:
-    metadata:
-      labels:
-        app: {{ .Values.image.label }}
-    spec:
-      imagePullSecrets:
-      - name: acrsecret
-      containers:
-      - image: "{{ .Values.repository.image }}:{{ .Values.repository.tag }}"
-        imagePullPolicy: {{ .Values.repository.pullPolicy }}
-        name: {{ .Values.image.name }}
-        ports:
-          - containerPort: {{ .Values.image.containerPort }}
-            name: http
-            protocol: TCP
-          - containerPort: 443
-            name: https
-            protocol: TCP
-        env:
-          - name: SQL_USER
-            valueFrom:
-              secretKeyRef:
-                name: sql
-                key: sql_user
-          - name: SQL_PASSWORD
-            valueFrom:
-              secretKeyRef:
-                name: sql
-                key: sql_password
-          - name: SQL_SERVER
-            valueFrom:
-              secretKeyRef:
-                name: sql
-                key: sql_server
-          - name: SQL_DBNAME
-            valueFrom:
-              secretKeyRef:
-                name: sql
-                key: sql_dbname
