--- apis/MyDriving.POIService.v2/helm/templates/ingress.yaml
@@ -1,35 +1,0 @@
-{{- if .Values.ingress.enabled -}}
-{{- $serviceName := include "chart.fullname" . -}}
-{{- $servicePort := .Values.service.externalPort -}}
-{{- $path := .Values.ingress.path -}}
-apiVersion: extensions/v1beta1
-kind: Ingress
-metadata:
-  name: {{ template "chart.fullname" . }}
-  labels:
-    app: {{ template "chart.name" . }}
-    chart: {{ .Chart.Name }}-{{ .Chart.Version | replace "+" "_" }}
-    release: {{ .Release.Name }}
-    heritage: {{ .Release.Service }}
-  annotations:
-    {{- range $key, $value := .Values.ingress.annotations }}
-      {{ $key }}: {{ $value | quote }}
-    {{- end }}
-spec:
-  rules:
-    {{- range $key, $value := .Values.ingress.rules }}
-    - host: {{ $key }}
-      http:
-        paths:
-        {{- range $path := . }}
-          - path: {{ $path.path }}
-            backend:
-              serviceName: {{ $path.serviceName }}
-              servicePort: {{ $path.servicePort }}
-        {{- end -}}
-    {{- end -}}
-  {{- if .Values.ingress.tls }}
-  tls:
-{{ toYaml .Values.ingress.tls | indent 4 }}
-  {{- end -}}
-{{- end -}}
