--- apis/MyDriving.POIService.v2/host.json
@@ -1,2 +1,0 @@
-{
-}
