--- apis/MyDriving.ServiceObjects/BaseDataObject.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace MyDriving.ServiceObjects
-{
-    public class BaseDataObject
-    {
-        public string Id { get; set; }
-
-        public BaseDataObject()
-        {
-            Id = Guid.NewGuid().ToString();
-        }
-    }
-}
