--- apis/MyDriving.ServiceObjects/POIType.cs
@@ -1,11 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for details.
-
-namespace MyDriving.ServiceObjects
-{
-    public enum POIType
-    {
-        HardAcceleration = 1,
-        HardBrake = 2
-    }
-}
