+++ apis/poi/MyDriving.POIService.v2/Dockerfile
@@ -1,0 +1,6 @@
+#FROM microsoft/azure-functions-runtime:2.0.0-jessie
+
+FROM azurefunctionstest/azure-functions-dotnet2.0:dev-jessie
+
+ENV AzureWebJobsScriptRoot=/home/site/wwwroot
+COPY ./bin/ /home/site/wwwroot
