+++ apis/poi/MyDriving.POIService.v2/helm/Chart.yaml
@@ -1,0 +1,5 @@
+apiVersion: v1
+appVersion: "1.0"
+description: A chart for deploying the My Driving POI API
+name: mydriving-poi
+version: 1.0.0
