+++ apis/poi/MyDriving.POIService.v2/helm/templates/service.yaml
@@ -1,0 +1,17 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: {{ template "chart.fullname" . }}
+spec:
+  type: ClusterIP
+  selector:
+    app: {{ .Values.image.label }}
+  ports:
+    - protocol: TCP
+      name: {{ .Values.image.name }}http
+      port: {{ .Values.service.externalPort }}
+      targetPort: {{ .Values.image.containerPort }}
+    - protocol: TCP
+      name: {{ .Values.image.name }}https
+      port: 443
+      targetPort: 443
