+++ apis/poi/MyDriving.POIService.v2/helm/values.yaml
@@ -1,0 +1,47 @@
+# Default values for helmv2.
+# This is a YAML-formatted file.
+# Declare variables to be passed into your templates.
+
+replicaCount: 2
+repository:
+  image: acrreplace.azurecr.io/devopsoh/imagereplace
+  tag: latest
+  pullPolicy: Always
+image:
+  name: mydrive-poi
+  label: mydrive-poi
+  containerPort: 8080
+service:
+  externalPort: 80
+ingress:
+  enabled: true
+  annotations:
+    kubernetes.io/ingress.class: traefik
+  rules:
+    akstraefikrg9aa1.eastus.cloudapp.azure.com:
+      - path: /api/GetAllPOIs
+        serviceName: api-poi-mydrive-poi
+        servicePort: 80
+
+  tls: []
+  #  - secretName: chart-example-tls
+  #    hosts:
+  #      - chart-example.local
+
+resources: {}
+  # We usually recommend not to specify default resources and to leave this as a conscious
+  # choice for the user. This also increases chances charts run on environments with little
+  # resources, such as Minikube. If you do want to specify resources, uncomment the following
+  # lines, adjust them as necessary, and remove the curly braces after 'resources:'.
+  # limits:
+  #  cpu: 100m
+  #  memory: 128Mi
+  # requests:
+  #  cpu: 100m
+  #  memory: 128Mi
+
+nodeSelector: {}
+
+tolerations: []
+
+affinity: {}
