+++ apis/poi/MyDriving.POIService.v2/host.json
@@ -1,0 +1,2 @@
+{
+}
