+++ apis/poi/MyDriving.ServiceObjects/BaseDataObject.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MyDriving.ServiceObjects
+{
+    public class BaseDataObject
+    {
+        public string Id { get; set; }
+
+        public BaseDataObject()
+        {
+            Id = Guid.NewGuid().ToString();
+        }
+    }
+}
